Cosolidate test_paradedb_queries and test_sql_generation into test_queries
